+++ packages/containers/psammead-timestamp-container/src/Utilities/timestampUtilities.test.js
@@ -4,7 +4,7 @@ import {
   formatUnixTimestamp,
   isValidDateTime,
 } from './timestampUtilities';
-import timestampGenerator from './helpers/testHelpers';
+import timestampGenerator from '../helpers/testHelpers';
 
 const timezone = 'Europe/London';
 const timestamp = 1539969006000; // 19 October 2018
