+++ packages/containers/psammead-timestamp-container/utilities/package.json
@@ -1,7 +1,7 @@
 {
     "name": "@bbc/psammead-timestamp-container/utilities",
     "private": true,
-    "main": "../dist/timestampUtilities",
-    "module": "../esm/timestampUtilities",
+    "main": "../dist/Utilities",
+    "module": "../esm/Utilities",
     "sideEffects": false
   }
